create make commands for tests
